Fix build: AGP 9.x + KSP 2.x + Hilt 2.60 compatibility
- ksp 2.2.10-1.0.29 -> 2.2.10-2.0.2 (new versioning scheme)
- Restore kotlin.android with android.builtInKotlin=false flag
- hilt 2.56 -> 2.60, firebase-perf 1.4.2 -> 2.0.2 (AGP 9.x Transform API)
- crashlytics 3.0.3 -> 3.0.7
- firebase-remote-config-ktx -> firebase-config-ktx (renamed artifact)
- Replace FirebaseFirestore.CACHE_SIZE_UNLIMITED with PersistentCacheSettings API
- Add missing kotlin.math.floor import in LayerCalculator
- Fix icon extension property imports (ArrowBack, Visibility, Warning, Info)
- Fix LayerPreset package: catalog -> model
- Add @file:OptIn(ExperimentalMaterial3Api::class) to affected screens
- Fix Vertex.x/y Double -> Float for Compose Offset

+++ app/src/main/java/com/example/cubemaster/presentation/geometry/GeometryScreen.kt
@@ -338,7 +338,7 @@ private fun PolygonPreview(
         fun toCanvas(v: Vertex): Offset {
             val x = padding + (v.x - minX) * scaleX + offset.x
             val y = size.height - padding - (v.y - minY) * scaleY + offset.y
-            return Offset(x, y)
+            return Offset(x.toFloat(), y.toFloat())
         }
 
         val path = Path()
